Added script.js and improved media queries
Added script to allow user to increment/decrement the number of episodes watched. Also edited the button size depending on device and removed the hover styling for mobile devices so that those no longer get stuck. Finally, I disabled the double-tap-to-zoom feature so that mobile users don't zoom in on their browser window when trying to interact with buttons.